Fix item model, nil package reference bug
As reported by Olaf here:

http://groups.google.com/group/munkiserver-dev/browse_thread/thread/e7ea86786b70556f

The issue sprouts from a relation created to a package branch within a particular unit.  At a later time, a package from that unit that belongs to that package branch is removed (or moved to another unit or environment), thus leaving the manifest (e.g., computer) with no available package.  The root issue will be fixed once package branches belong to a specific unit.

In the meantime, I've added a rake task called "chore:destroy_stale_item_records" to remove any records with references to a nil package.

I've also updated the view to not display those stale item records.  Stale item records will be removed from a manifest if the manifest is edited and saved.