Fix #145 by creating new action for munki client API
Added a new action: ComputersController#show_plist. This is a somewhat confusing name because you can actually ask for a plist version of ComputersController#show. The purpose of this action is to serve the munki client API when retrieving a computer manifest.

Eventually, I'd like to split the munki API into it's own controller, or something more clear than this. But for now, this will do.