Add rake to destroy unused package branches
An "unused package branch" is a package branch with no packages. This was created because manifests were referencing package branches (through item join models) that had no packages. When it came time to show a plist or HTML version of the resource, the lack of a package to display caused issues.

This task will fix a bad state, but won't stop the state from happening. If this continues to occur, I will have to investigate what causes the state and stop that.

Until then, this'll do.