Add fallback manifest (closes #146)
As suggested by @mkuron, added a route matching http://localhost/site_default, which is the last URL a munki client tries if it has no ClientIdentifier set.

How to use this feature:
Make a computer record with a MAC address of 00:00:00:00:00:00 and assign any pre/post flight or bootstrapping packages to configure the computer. This manifest will be served to EVERY computer that does not have a ClientIdentifier set.